contracts: settlement waterfall + SC-FP-008 loss waterfall
Unblocked by the SPEC-04 wiring.

JobVault.acceptDelivery (SC-JV-005/009) - 'the fall':
- customer-only, Delivered -> Accepted, nonReentrant, status set before any transfer
- reads openAdvanceOf, approves the pool for exactly principal+fee, pool pulls
  inside repayAdvance, THEN the remainder goes to the operator
- seniority is control flow, not convention (ADR-010). MockUSDC logs every
  movement and test_acceptDelivery_poolIsPaidBeforeOperator asserts the pool's
  transfer index precedes the operator's

JobVault.refund/cancel (SC-JV-006/010):
- refund: operator/admin any time, customer once the deadline passes. Customer
  is made whole FIRST, then FloatPool.writeOff absorbs the loss
- cancel: unfunded jobs only; no advance can exist pre-funding per SC-FP-001

FloatPool.repayAdvance: exact-amount only, splits 20% of the fee to the reserve
(SC-FP-005), increments acceptedJobs - the on-chain half of the flywheel.

FloatPool.writeOff: SC-FP-008 stages bond -> reserve -> LP shares with an event
each. The bond stage emits 0 rather than being skipped, so the sequence shape is
stable for when the P1 bond lands. Losses socialize by reducing totalAssets;
share count is untouched, so LP positions stay proportional.

One lint suppression added with justification: block.timestamp in the refund
deadline check. A deadline is inherently a wall-clock comparison and validator
drift is seconds against a multi-day window.

102 tests green, zero build warnings.

+++ contracts/test/Advance.t.sol
@@ -298,10 +298,10 @@ contract AdvanceTest is Test {
     function test_utilizationCap_stopsGlobalLendingAtEightyPercent() public {
         // 10 orgs × 10% each would be 100% utilization; the global cap must bite at 80%.
         uint256 issued;
-        for (uint256 i = 1; i <= 10; i++) {
-            address org = address(uint160(0x1000 + i));
+        for (uint160 i = 1; i <= 10; i++) {
+            address org = address(0x1000 + i);
             bytes32 job = keccak256(abi.encode("job", i));
-            address cust = address(uint160(0x2000 + i));
+            address cust = address(0x2000 + i);
 
             usdc.mint(cust, 30_000_000);
             vm.prank(cust);
